Add types to index config, apply type filter to all queries (#261)
Both resources and items are stored in the same index. With the
previous data, we did not notice the following issue: for a query
over all fields (with `q=X`), we can get lobid-items where we
expect lobid-resources. When processing an item as a resource, we
get errors, manifested as the issue @literarymachine reported in
https://github.com/lobid/lodmill/pull/261#issuecomment-36107963